projects/techmantra/requirements.txt
# Core app
streamlit==1.56.0
pandas==2.3.3
python-dotenv==1.2.2

# AI + NLP
ollama==0.6.1
medspacy==1.3.1
scispacy==0.6.2
sentence-transformers==5.4.1
chromadb==1.5.8
langchain-text-splitters==1.1.2

# Integrations
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1

# Voice input
streamlit_mic_recorder==0.0.8

# Testing / demo scripts
pytest==9.0.3
